backend/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==4.0.3
attrs==25.1.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
exceptiongroup==1.2.2
Flask==3.1.0
Flask-Cors==5.0.0
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.18
langchain-community==0.3.17
langchain-core==0.3.34
langchain-openai==0.3.4
langchain-postgres==0.0.13
langchain-text-splitters==0.3.6
langsmith==0.3.7
MarkupSafe==3.0.2
marshmallow==3.26.1
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.61.1
orjson==3.10.15
packaging==24.2
pgvector==0.3.6
propcache==0.2.1
psycopg==3.2.4
psycopg-binary==3.2.4
psycopg-pool==3.2.4
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pypdf==5.2.0
PyPDF2==3.0.1
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.38
tenacity==9.0.0
tiktoken==0.8.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
yarl==1.18.3
zstandard==0.23.0
